refactor: move logic to the increase dissolve delay modal and chore
